app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["moonpeter <moonpeter@kakao.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
dj-rest-auth = "^2.2.3"
django-allauth = "^0.49.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
debugpy = "^1.5.1"
google-auth = "^2.6.2"
Pillow = "^9.0.1"
sentry-sdk = "^1.5.8"
pytest-cov = "^3.0.0"
django-bakery = "^0.12.7"
pytest-django = "^4.5.2"
django-factory_boy = "^1.0.0"
boto3 = "^1.21.40"
django-storages = "^1.12.3"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
black = "^22.1.0"
coverage = "^6.3.2"
debugpy = "^1.5.1"
schema = "^0.7.5"
django-extensions = "^3.1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
